requirements.txt
aniso8601==7.0.0
asgiref==3.2.3
boto3==1.12.2
botocore==1.15.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
django-graphql-geojson==0.1.4
django-graphql-jwt==0.3.0
django-storages==1.9.1
django-versatileimagefield==2.0
docutils==0.15.2
graphene==2.1.8
graphene-django==2.8.2
graphql-core==2.2.1
graphql-relay==2.0.1
gunicorn==20.0.4
idna==2.8
jmespath==0.9.4
Pillow==6.2.1
Pint==0.10.1
pip-chill==0.1.8
promise==2.2.1
psycopg2==2.8.4
pycrypto==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2019.3
requests==2.22.0
Rx==1.6.1
s3transfer==0.3.3
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
